- D3D9Client x64 build
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@1438 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/CloudMgr.cpp
@@ -24,7 +24,7 @@ using namespace oapi;
 CloudManager::CloudManager(D3D9Client *gclient, const vPlanet *vplanet)
 : TileManager (gclient, vplanet)
 {
-	int len = strlen(objname)+7;
+	size_t len = strlen(objname)+7;
 	char *texname = new char[len];
 	strcpy_s(texname, len, objname);
 	strcat_s(texname, len, "_cloud");
